requirements.txt
Django<1.9
pytz
django-bootstrap3
django-bootstrap3-datepicker
djangorestframework<3.7
